refactor: feature app -> application
