+++ src/app/(list)/category/[category]/_components/PostCardList.tsx
@@ -4,7 +4,7 @@ import React from 'react'
 import PostCard from '@/feature/post/components/PostCard'
 import CategoryChip from '@/feature/post/components/CategoryChip'
 import usePagination from '@/components/Pagination/usePagination'
-import AppFeature from '@/feature/app'
+import AppFeature from '@/feature/application'
 import * as style from 'src/feature/post/components/style/cardListFrame.css'
 import Link from 'next/link'
 import { PostFrontMatter } from '@/entity/post/type'
